requirements.txt
asgiref==3.2.10
boto3==1.17.62
botocore==1.20.62
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-storages==1.11.1
gunicorn==20.0.4
Jinja2==2.10.1
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==6.2.2
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2018.5
s3transfer==0.4.2
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.4
whitenoise==5.1.0
